Allow to have lectures in the interface, ie non-interactive exercises Use it for the lesson chooser instead of a separate frame
This patch is a bit intrusive, sorry about it, but adding a superclass
to one of the most overloaded class of our hierarchie wasn't easy. I
should have split it in parts, in a better world, but life is crap
sometimes.

The good point is that this feature will help solving a lot of issues,
such as slow loading (load only one lesson at a time), or difficulties
for teacher to cume up with their own lessons.