Create a Entity.runIt() method
It's in charge of running the entity, which is not a trivial task as
it depends on the universe (bat entities are not run as the other
ones) and on the language (java entities are just run by launching
their run() method that was written by the student; python entities
are run by injecting the code written by the student into a Scripting
engine; and Lightbot entities are run by executing their run() method
that interprets the code "written" by the student).

Also, improve a bit the documentation in this area of the code.