Kill a rogue call to mutateEntity by simplifying BDR2
We want to check whether the dance floor indications correctly lead to
the same trajectory for each dancer, so the entity does a bit more
than just dancing.

But now that we have world.getParam() (that was introduced with
TurtleWorlds), we don't have to hack the JLM internal directly this way.

This simplification will ease further refactoring of mutateEntity()