Building a Mac .app bundle
All the files necessary to building a Mac application bundle. The ant
tasks are
-- build-Mac.app-bundle: packs the various files into the .app bundle
(jlmxxx.jar is assumed built)
-- Mac.app: bulds the main jar and packs it into a bundle

The icons are what they are, none too nice especially the 16x16 size
one. Is there someone out there who knows how to draw nice icons
willing to make a set for this project?