Change the world so that the entity don't have to declare Entity copy()
That method had to be declared in every single Entity child classes,
and it was used on one location only (the reset of one world to the
state of another world).

So change that location so that it uses the "void copy(Entity)" method
instead.

This should simplify a bit the way we build new universes.