Improve a sentence
One could understand that the "it is often so in CS" refers to "it is
disturbing" while I meant it to refer to the fact that the first line
is #0 and so on. I fail to rephrase it correctly, so I just change
"disturbing" to "surprising", with which the misunderstanding is ok.