one thread less in ExerciseRunner
There is no need to start a thread (out of a CachedThreadPool) just to
start the threads that will do the work. Instead, do it directly from
the calling thread.

This should not have any impact on the performance, since
World::runEntities() (the most expensive here) does just start one
thread per entity.