Move lessons and exercises back from webPLM to PLM
We want webPLM to be a "dumb" frontend on top of PLM. All the core logic
would reside in PLM. In particular lesson and exercise loading and
management should be handled by PLM.

There were a bunch of places in PLM where files were loaded from the
filesystem instead of the classpath. This commit also fixes that and
only loads files from the classpath. This allows the webPLM tests to
pass even if not run from the source directory.

Eventually we might want to load exercices and lessons from a separate
repository and possibly look them up by commit hash. But this requires a
careful design phase that we haven't gone through yet.