Abstract over how files and classes are loaded.
Make FileUtils a class parameterized by a classloader and parameterize
every former static use of FileUtils in PLM. Also pass the classloader
as an argument everywhere it was used statically before. This should
allow us to fix the tests (not done in that commit) and eventually I
hope to make "sbt run" work. Also this will allow us to read the files
from the network if we ever want to.