Let's first make sure that one lesson passes
PLM is about to take fire and burn to the ground. With this commit, I
reduce the code surface in a last attempt to get things under control.

My goal is to make sure that lesson "welcome" works well, both
remotely and through the server, and with Java/Scala/Python.

Then I'll simplify the codebase in a way that these feature remain.
These simplifications may require rather intrusive changes in the
javascript world.

Automatic testing of the javascript UI would be really welcome too.

Features put on hold until after this stabilization/simplification:
 - Any other universes, from other lessons
 - Compilation in separated Judges (Speedup work reduced the
   sandboxing that should now be reinforced)
 - C Lang (is seriously broken since webPLM anyway)
 - Blockly language (may still work, but it's hard to test w/o
   automatic testing of javascript)